move QrReaderScannerStyles
