+++ react/src/components/QrReader/components/QrReaderScanner.tsx
@@ -8,7 +8,7 @@ import * as React from "react";
 import { MutableRefObject, useEffect, useRef } from "react";
 import { BrowserQRCodeReader, IScannerControls } from "@zxing/browser";
 import { Result } from "@zxing/library";
-import { styles } from "../styles";
+import { styles } from "./QrReaderScannerStyles";
 
 export type OnResultFunction = (
   /**
