+++ react/src/components/QrReader/components/QrReaderScannerStyles.ts
